Habitat Data

Datasets used to calculate spawning, instream, and floodplain rearing habitat areas for Chinook Salmon and Steelhead

Instream ({watershedname}_instream)

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - flow value in cubic feet per second

FR_spawn_wua - Spawning WUA in square feet per 1000 feet

FR_fry_wua - Fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua - Juvenile WUA in square feet per 1000 feet

watershed - Name of watershed

Floodplain ({watershedname}_floodplain)

Datasets containing the total inuncated area in acres as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - Flow value in cubic feet per second

FR_floodplain_acres - Total inundated area in acres for Fall Run

SR_floodplain_acres - Total inundated area in acres for Spring Run

ST_floodplain_acres - Total inundated area in acres for Steelhead

watershed - Name of watershed

Modeling Details for Streams

American River

Antelope Creek

Battle Creek

Bear Creek

Bear River

Big Chico Creek

Butte Creek

Calaveras River

Clear Creek

Cosumnes River

Cottonwood Creek

Cow Creek

Deer Creek

Elder Creek

Feather River

Merced River

Mill Creek

Mokelumne River

Paynes Creek

Sacramento River

San Joaquin River

Stanislaus River

Stony Creek

Thomes Creek

Tuolumne River

Yuba River

Bypasses

Sutter and Yolo Bypasses

Delta

North and South Delta

Examples

american_river_instream
#> # A tibble: 72 × 6
#>    flow_cfs FR_spawn_sqm FR_fry_sqm FR_juv_sqm ST_spawn_sqm watershed     
#>       <dbl>        <dbl>      <dbl>      <dbl>        <dbl> <chr>         
#>  1      500      128256.    322492.    322492.      111239. American River
#>  2      800          NA     322086.    322086.          NA  American River
#>  3     1000      162239.    326565.    326565.      141084. American River
#>  4     1250          NA     329523.    329523.          NA  American River
#>  5     1500          NA     333709.    333709.          NA  American River
#>  6     1750          NA     336677.    336677.          NA  American River
#>  7     2000      171153.    339811.    339811.      143101. American River
#>  8     2750          NA     345352.    345352.          NA  American River
#>  9     3500          NA     307132.    307132.          NA  American River
#> 10     4500          NA     246900.    246900.          NA  American River
#> # ℹ 62 more rows
american_river_floodplain
#> # A tibble: 21 × 4
#>    flow_cfs FR_floodplain_acres ST_floodplain_acres watershed     
#>       <dbl>               <dbl>               <dbl> <chr>         
#>  1      500                0                   0    American River
#>  2      800                0                   0    American River
#>  3     1000                0                   0    American River
#>  4     1250                0                   0    American River
#>  5     1500                0                   0    American River
#>  6     1750                0                   0    American River
#>  7     2000                0                   0    American River
#>  8     2750                6.55                6.55 American River
#>  9     3500               22.0                22.0  American River
#> 10     4500               42.5                42.5  American River
#> # ℹ 11 more rows