Juvenile Inchannel Habitat

The 1980-1999 juvenile suitable inchannel rearing habitat area in square meters for Steelhead and 1980-2000 juvenile suitable inchannel rearing habitat area in square meters for Fall Run, Spring Run and Winter Run.

fr_juv

wr_juv

sr_juv

st_juv

lfr_juv

Format

Steelhead: 3 dimensional array [31 watersheds, 12 months, 20 years]

Fall Run, Late Fall Run, Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 21 years]

Details

Values created using the function set_instream_habitat