Habitat Modeling Status

This table summaries modeling availability based on species and lifestage combinations for each watershed

modeling_exist

Format

dataframe with 32 rows and 23 variables

Order

integer value representing watershed order in SIT model

Watershed

name of watershed

FR_spawn

TRUE if spawning habitat modeling exists for Fall Run

FR_fry

TRUE if fry rearing habitat modeling exists for Fall Run

FR_juv

TRUE if juvenile rearing habitat modeling exists for Fall Run

FR_floodplain

TRUE if floodplain rearing modeling exists for Fall Run

SR_spawn

TRUE if spawning habitat modeling exists for Spring Run

SR_fry

TRUE if fry rearing habitat modeling exists for Spring Run

SR_juv

TRUE if juvenile rearing habitat modeling exists for Spring Run

SR_floodplain

TRUE if floodplain rearing modeling exists for Spring Run

ST_spawn

TRUE if spawning habitat modeling exists for Steelhead

ST_fry

TRUE if fry rearing habitat modeling exists for Steelhead

ST_juv

TRUE if juvenile rearing habitat modeling exists for Steelhead

ST_floodplain

TRUE if floodplain rearing modeling exists for Steelhead

ST_adult

TRUE if adult rearing habitat modeling exists for Steelhead

LFR_spawn

TRUE if spawning habitat modeling exists for Late-fall Run

LFR_fry

TRUE if fry rearing habitat modeling exists for Late-fall Run

LFR_juv

TRUE if juvenile rearing habitat modeling exists for Late-fall Run

LFR_floodplain

TRUE if floodplain rearing modeling exists for Late-fall Run

WR_spawn

TRUE if spawning habitat modeling exists for Winter Run

WR_fry

TRUE if fry rearing habitat modeling exists for Winter Run

WR_juv

TRUE if juvenile rearing habitat modeling exists for Winter Run

WR_floodplain

TRUE if floodplain rearing modeling exists for Winter Run

Source

Sadie Gill sgill@flowwest.com

Examples

modeling_exist

#> # A tibble: 32 x 23
#>    Order Watershed   FR_spawn FR_fry FR_juv FR_floodplain SR_spawn SR_fry SR_juv
#>    <dbl> <chr>       <lgl>    <lgl>  <lgl>  <lgl>         <lgl>    <lgl>  <lgl> 
#>  1     1 Upper Sacr~ TRUE     FALSE  FALSE  FALSE         FALSE    FALSE  FALSE 
#>  2     2 Antelope C~ FALSE    FALSE  FALSE  FALSE         FALSE    FALSE  FALSE 
#>  3     3 Battle Cre~ TRUE     TRUE   TRUE   FALSE         TRUE     TRUE   TRUE  
#>  4     4 Bear Creek  FALSE    FALSE  FALSE  FALSE         NA       NA     NA    
#>  5     5 Big Chico ~ FALSE    FALSE  FALSE  TRUE          FALSE    FALSE  FALSE 
#>  6     6 Butte Creek FALSE    FALSE  FALSE  TRUE          TRUE     TRUE   TRUE  
#>  7     7 Clear Creek TRUE     TRUE   TRUE   FALSE         TRUE     TRUE   TRUE  
#>  8     8 Cottonwood~ TRUE     TRUE   TRUE   TRUE          FALSE    FALSE  FALSE 
#>  9     9 Cow Creek   FALSE    TRUE   TRUE   FALSE         NA       NA     NA    
#> 10    10 Deer Creek  FALSE    FALSE  FALSE  TRUE          FALSE    FALSE  FALSE 
#> # ... with 22 more rows, and 14 more variables: SR_floodplain <lgl>,
#> #   ST_spawn <lgl>, ST_fry <lgl>, ST_juv <lgl>, ST_floodplain <lgl>,
#> #   ST_adult <lgl>, LFR_spawn <lgl>, LFR_fry <lgl>, LFR_juv <lgl>,
#> #   LFR_floodplain <lgl>, WR_spawn <lgl>, WR_fry <lgl>, WR_juv <lgl>,
#> #   WR_floodplain <lgl>