Set Instream Habitat Area

This function returns an estimated instream habitat area based on watershed, species, life stage, and flow.

set_instream_habitat(watershed, species, life_stage, flow, ...)

Arguments

watershed

one of the watersheds defined for the SIT model

species

one of 'fr' (Fall Run), 'lfr' (Late Fall Run), sr' (Spring Run), or 'st' (Steelhead)

life_stage

life stage of fish, one of 'juv', 'adult' or 'fry'

flow

value used to determine habitat area

Value

habitat area in square meters

Details

The function relies on a dataframe called modeling_exist that contains data on whether the species is present in a watershed and whether habitat modeling exists. If a model for the watershed does exist, the function looks up the flow to weighted usable area (WUA) relationship (e.g. battle_creek_instream) and selects the correct WUA for the given flow, species, and life stage. This WUA is then multiplied by the watershed's typical rearing habitat extent length (stored in watershed_lengths), to return an estimate of suitable rearing habitat within the watershed. When additional species modeling is not available, the fall run WUA values are used (lengths are modified if the habitat extent varies across species). Also, if there is no modeling specifically for fry, then the juvenile value is used.

Lower-mid Sacramento River

The Lower-mid Sacramento River has two nodes, one above Fremont Weir (C134) and one below (C160). When calculating habitat for the Lower-Mid Sacramento river, calculate the habitat at each flow node and sum them proportion to the length of stream above and below the weir:

35.6/58 * (habitat at C134) + 22.4/58 * (habitat at C160)

Regional Approximation: When a watershed has no associated flow to WUA reltionship, a regional approximation is made. First, the mean WUA at the given flow value from a set of similar modeled watersheds nearby is calculated. Then the mean WUA is multiplied by the river length of the watershed of interest.

Below are the regions (defined by the downstream watershed) that contain watersheds with unmodeled spawning relationships. The modeled watersheds used to approximate spawning area for the unmodeled watersheds are marked with an asterisk.

Upper-mid Sacramento River

Battle Creek*

Bear Creek

Big Chico Creek

Butte Creek*

Clear Creek*

Cottonwood Creek*

Cow Creek*

Deer Creek

Elder Creek

Mill Creek

Paynes Creek

Stony Creek

Thomes Creek

South Delta

Calaveras River*

Cosumnes River

Mokelumne River*

Examples

# Fry rearing habitat value in square meters for Fall Run in the Merced River at 425 cfs.
set_instream_habitat('Merced River', 'fr', 'fry', 425) # habitat modeling exists
#> [1] 446380.6
# Juvenile rearing habitat value in square meters for Fall Run in Elder Creek at 300 cfs.
set_instream_habitat('Elder Creek', 'fr', 'juv', 300) # no habitat modeling exists, composite used
#> [1] 31964.76