Habitat Extent Lengths

A dataset containing the length of rearing and spawning extent within each CVPIA watershed for different species.

watershed_lengths

Format

dataframe with 57 rows and 7 variables

order

integer value representing watershed order in SIT model

watershed

name of watershed

lifestage

habitat type by lifestage, 'spawing' or 'rearing'

miles

length in miles

feet

length in feet

source

Expert who delineated habitat extents

species

species of habitat extent

Source

Salmonid Extents

Details

Information comes from expert outreach undertaken in 2017. Updated in 2021 to include late fall run extents from expert outreach.

Examples

watershed_lengths
#> # A tibble: 168 x 8
#>    order watershed            lifestage miles   feet source species date_updated
#>    <dbl> <chr>                <chr>     <dbl>  <dbl> <chr>  <chr>   <chr>       
#>  1     0 Wilkins Slough  to ~ NA         38.2 2.02e5 NA     fr      6/1/2017    
#>  2     0 Feather to American  NA         19.7 1.04e5 NA     fr      6/1/2017    
#>  3     0 Lower-mid Sac above~ NA         35.6 1.88e5 Mike ~ fr      6/1/2017    
#>  4     0 Lower-mid Sac below~ NA         22.4 1.18e5 Mike ~ fr      6/1/2017    
#>  5     0 Sacramento River (B~ NA        189.  9.98e5 NA     fr      6/1/2017    
#>  6     0 Sacramento River (F~ NA         33.4 1.77e5 NA     fr      6/1/2017    
#>  7     0 Sacramento River (k~ NA         55.5 2.93e5 NA     fr      6/1/2017    
#>  8     0 wilkins slough  to ~ NA         38.2 2.02e5 NA     fr      6/1/2017    
#>  9     1 Upper Sacramento Ri~ spawning   59.4 3.13e5 Jim S~ fr      6/1/2017    
#> 10     1 Upper Sacramento Ri~ rearing    59.2 3.13e5 Jim S~ fr      6/1/2017    
#> # ... with 158 more rows