Bypass Flow to Habitat Area Relationships

A dataset containing the suitable habitat area in square meters as a function of flow in cubic feet per second.

Source

Sutter and Yolo Bypass Modeling Details

yolo_bypass_habitat

flow_cfs

integer flow value in cubic feet per second

Yolo Bypass 1

suitable rearing area in square meters in the Yolo Bypass, Fremont Weir to Sacramento Weir

Yolo Bypass 2

suitable rearing area in square meters in the Yolo Bypass below Sacramento Weir

sutter_bypass_habitat

flow_cfs

integer flow value in cubic feet per second

Sutter Bypass 1

suitable rearing area in square meters in the Sutter Bypass, to Moulton Weir

Sutter Bypass 2

suitable rearing area in square meters in the Sutter Bypass, to Colusa Weir

Sutter Bypass 3

suitable rearing area in square meters in the Sutter Bypass, to Tisdale Weir

Sutter Bypass 4

suitable rearing area in square meters in the Sutter Bypass below Tisdale Weir

Examples

yolo_bypass_habitat

#> # A tibble: 41 x 3
#>    flow_cfs `Yolo Bypass 1` `Yolo Bypass 2`
#>       <dbl>           <dbl>           <dbl>
#>  1       10          27932.         253808.
#>  2       20         307501.         702708.
#>  3       30         425135.         905245.
#>  4       40         521795.         971859.
#>  5       50         600826.        1044463.
#>  6       60         642940.        1136821.
#>  7       70         640828.        1096337.
#>  8       80         606158.        1202895.
#>  9       90         622189.        1292838.
#> 10      100         593637.        1375864.
#> # ... with 31 more rows