Habitat Data

Datasets used to calculate spawning, instream, and floodplain rearing habitat areas for Chinook Salmon and Steelhead

Instream ({watershedname}_instream)

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - flow value in cubic feet per second

FR_spawn_wua - Spawning WUA in square feet per 1000 feet

FR_fry_wua - Fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua - Juvenile WUA in square feet per 1000 feet

watershed - Name of watershed

Floodplain ({watershedname}_floodplain)

Datasets containing the total inuncated area in acres as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - Flow value in cubic feet per second

FR_floodplain_acres - Total inundated area in acres for Fall Run

SR_floodplain_acres - Total inundated area in acres for Spring Run

ST_floodplain_acres - Total inundated area in acres for Steelhead

watershed - Name of watershed

Modeling Details for Streams

American River

Antelope Creek

Battle Creek

Bear Creek

Bear River

Big Chico Creek

Butte Creek

Calaveras River

Clear Creek

Cosumnes River

Cottonwood Creek

Cow Creek

Deer Creek

Elder Creek

Feather River

Merced River

Mill Creek

Mokelumne River

Paynes Creek

Sacramento River

San Joaquin River

Stanislaus River

Stony Creek

Thomes Creek

Tuolumne River

Yuba River

Bypasses

Sutter and Yolo Bypasses

Delta

North and South Delta

Examples

american_river_instream
#> # A tibble: 61 × 6
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua ST_spawn_wua watershed     
#>       <dbl>        <dbl>      <dbl>      <dbl>        <dbl> <chr>         
#>  1        0           0         NA         NA            0  American River
#>  2       50       13209.        NA         NA        11332. American River
#>  3      100       25142.        NA         NA        21601. American River
#>  4      150       35891.        NA         NA        30880. American River
#>  5      200       45542.        NA         NA        39239. American River
#>  6      300       61891.     59897     112381.       53466. American River
#>  7      400       74822.        NA         NA        64789. American River
#>  8      500       85255.     54065     108736.       73943. American River
#>  9      750          NA      49959.     99007           NA  American River
#> 10     1000      107844.     42334      93968.       93782. American River
#> # ℹ 51 more rows
american_river_floodplain
#>    flow_cfs FR_floodplain_acres ST_floodplain_acres      watershed
#> 1      1200                0.00                0.00 American River
#> 2      2200                6.00                6.00 American River
#> 3      3200              159.88              159.88 American River
#> 4      4200              228.01              228.01 American River
#> 5      5200              268.53              268.53 American River
#> 6      5570              294.31              294.31 American River
#> 7      6200              338.22              338.22 American River
#> 8      7200              390.14              390.14 American River
#> 9      8200              454.77              454.77 American River
#> 10     9200              499.83              499.83 American River
#> 11    10200              534.65              534.65 American River
#> 12    11200              568.76              568.76 American River
#> 13    12200              597.86              597.86 American River
#> 14    13200              622.60              622.60 American River
#> 15    14200              647.06              647.06 American River
#> 16    15200              672.88              672.88 American River
#> 17    16200              704.83              704.83 American River
#> 18    17200              729.66              729.66 American River
#> 19    18200              750.91              750.91 American River
#> 20    19200              776.30              776.30 American River
#> 21    20200              800.17              800.17 American River
#> 22    21200              826.44              826.44 American River
#> 23    22200              846.74              846.74 American River
#> 24    23200              862.57              862.57 American River
#> 25    24200              880.66              880.66 American River
#> 26    25200              900.47              900.47 American River
#> 27    26200              919.24              919.24 American River
#> 28    27200              937.71              937.71 American River
#> 29    28200              953.54              953.54 American River
#> 30    29200              969.11              969.11 American River
#> 31    30200              984.67              984.67 American River
#> 32    31200             1000.23             1000.23 American River
#> 33    32200             1016.32             1016.32 American River
#> 34    33200             1032.41             1032.41 American River