Habitat Data

Datasets used to calculate spawning, instream, and floodplain rearing habitat areas for Chinook Salmon and Steelhead

Instream ({watershedname}_instream)

Datasets containing the Weighted Usable Area (WUA) in square feet per 1000 feet as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - flow value in cubic feet per second

FR_spawn_wua - Spawning WUA in square feet per 1000 feet

FR_fry_wua - Fry (up to 50 mm) WUA in square feet per 1000 feet

FR_juv_wua - Juvenile WUA in square feet per 1000 feet

watershed - Name of watershed

Floodplain ({watershedname}_floodplain)

Datasets containing the total inuncated area in acres as a function of flow in cubic feet per second for Fall Run (FR), Spring Run (SR), Winter Run (WR), and Steelhead (ST) depending on species occurance and modeling availability.

Typical data structure for streams:

flow_cfs - Flow value in cubic feet per second

FR_floodplain_acres - Total inundated area in acres for Fall Run

SR_floodplain_acres - Total inundated area in acres for Spring Run

ST_floodplain_acres - Total inundated area in acres for Steelhead

watershed - Name of watershed

Modeling Details for Streams

American River

Antelope Creek

Battle Creek

Bear Creek

Bear River

Big Chico Creek

Butte Creek

Calaveras River

Clear Creek

Cosumnes River

Cottonwood Creek

Cow Creek

Deer Creek

Elder Creek

Feather River

Merced River

Mill Creek

Mokelumne River

Paynes Creek

Sacramento River

San Joaquin River

Stanislaus River

Stony Creek

Thomes Creek

Tuolumne River

Yuba River

Bypasses

Sutter and Yolo Bypasses

Delta

North and South Delta

Examples

american_river_instream

#> # A tibble: 61 x 6
#>    flow_cfs FR_spawn_wua FR_fry_wua FR_juv_wua ST_spawn_wua watershed     
#>       <dbl>        <dbl>      <dbl>      <dbl>        <dbl> <chr>         
#>  1        0           0         NA         NA            0  American River
#>  2       50       13209.        NA         NA        11332. American River
#>  3      100       25142.        NA         NA        21601. American River
#>  4      150       35891.        NA         NA        30880. American River
#>  5      200       45542.        NA         NA        39239. American River
#>  6      300       61891.     59897     112381.       53466. American River
#>  7      400       74822.        NA         NA        64789. American River
#>  8      500       85255.     54065     108736.       73943. American River
#>  9      750          NA      49959.     99007           NA  American River
#> 10     1000      107844.     42334      93968.       93782. American River
#> # ... with 51 more rows

american_river_floodplain

#> # A tibble: 34 x 4
#>    flow_cfs FR_floodplain_acres ST_floodplain_acres watershed     
#>       <dbl>               <dbl>               <dbl> <chr>         
#>  1     1200                0                   0    American River
#>  2     2200                6.00                6.00 American River
#>  3     3200              160.                160.   American River
#>  4     4200              228.                228.   American River
#>  5     5200              269.                269.   American River
#>  6     5570              294.                294.   American River
#>  7     6200              338.                338.   American River
#>  8     7200              390.                390.   American River
#>  9     8200              455.                455.   American River
#> 10     9200              500.                500.   American River
#> # ... with 24 more rows