Habitat Extent Lengths

A dataset containing the length of rearing and spawning extent within each CVPIA watershed for different species.

watershed_lengths

Format

dataframe with 57 rows and 7 variables

order

integer value representing watershed order in SIT model

watershed

name of watershed

lifestage

habitat type by lifestage, 'spawing' or 'rearing'

miles

length in miles

feet

length in feet

source

Expert who delineated habitat extents

species

species of habitat extent

Source

FlowWest, add url to shp file when mapping is complete

Details

Information comes from expert outreach undertaken in 2017.

Examples

watershed_lengths

#> # A tibble: 94 x 7
#>    order watershed              lifestage miles    feet source          species
#>    <dbl> <chr>                  <chr>     <dbl>   <dbl> <chr>           <chr>  
#>  1     1 Upper Sacramento River spawning  59.4  313369. Jim Smith       fr     
#>  2     1 Upper Sacramento River rearing   59.2  312500. Jim Smith       fr     
#>  3     2 Antelope Creek         spawning  23.7  125060. Matt Johnson    fr     
#>  4     2 Antelope Creek         rearing   31.6  166949. Matt Johnson    fr     
#>  5     3 Battle Creek           spawning   3.97  20945. Jim Smith       fr     
#>  6     3 Battle Creek           rearing    5.87  31003. Jim Smith       fr     
#>  7     4 Bear Creek             spawning  17.1   90309. Tricia Bratcher fr     
#>  8     4 Bear Creek             rearing   17.3   91324. Tricia Bratcher fr     
#>  9     5 Big Chico Creek        rearing   15.1   79488. Clint Garman    fr     
#> 10     5 Big Chico Creek        spawning  15.1   79488. Clint Garman    fr     
#> # ... with 84 more rows